+++ enclone_paper/src/lib.rs
@@ -1,1 +1,3 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
+
+pub mod public;
